Refactor: Clean up PR, separate tests, and update gitignore
