--- Scripts/data_reconciliation.log
@@ -1,16 +1,0 @@
-2025-07-22 16:11:59,750 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 16:11:59,751 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 16:11:59,751 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 16:11:59,751 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 16:11:59,751 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 16:11:59,755 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 16:11:59,755 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 16:11:59,755 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 16:11:59,763 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 16:11:59,763 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 16:11:59,763 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 16:11:59,763 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 16:11:59,764 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 16:11:59,764 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 16:11:59,764 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 16:11:59,764 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
