--- Scripts/explanation.txt
@@ -1,18 +1,0 @@
-# Data Imputation and Reconciliation Task Explanation
-
-This task implements data cleaning and preprocessing for sensor data in the DVR (Data Validation and Reconciliation) project. The implementation is in `data_reconciler.py` and processes sensor data from `sensor_data.parquet`.
-
-### What Was Done
-1. **Data Loading**: Loaded sensor data from `sensor_data.parquet` using Pandas, with columns like `id`, `timestamp`, `sensor1`, and `sensor2`.
-2. **Imputation**: Filled missing (`NaN`) values in numeric sensor columns using Linear Interpolation to estimate values based on neighboring data points. Remaining `NaN` values, if any, were filled with the column mean using `SimpleImputer`.
-3. **Smoothing**: Applied Exponential Moving Average (EMA) with `span=3` to reduce noise in sensor data, preserving `id` and `timestamp` columns.
-4. **Logging**: Logged all imputation and smoothing operations (e.g., imputed values, smoothed values) to `data_reconciliation.log` for transparency.
-5. **Output**: Saved cleaned data to `cleaned_data.parquet` for further analysis.
-6. **Testing**: Included unit tests using `unittest` to verify imputation and smoothing accuracy, ensuring no `NaN` values remain and `id`/`timestamp` columns are unchanged.
-
-### Files
-- `data_reconciler.py`: Contains the `DataReconciler` class and unit tests.
-- `cleaned_data.parquet`: Cleaned sensor dataset.
-- `data_reconciliation.log`: Log of processing steps.
-
-The implementation ensures robust data cleaning with detailed logging and testing, making the sensor data reliable for downstream tasks in the DVR project.
