--- data_reconciliation.log
@@ -1,64 +1,0 @@
-2025-07-22 15:18:48,791 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:18:48,791 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:18:48,792 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:18:48,792 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:18:48,793 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:18:48,794 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:18:48,794 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:18:48,794 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 15:18:48,817 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:18:48,817 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:18:48,818 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:18:48,818 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:18:48,820 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:18:48,820 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:18:48,821 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:18:48,821 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 15:18:52,333 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:18:52,333 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:18:52,333 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:18:52,333 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:18:52,334 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:18:52,334 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:18:52,334 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:18:52,334 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 15:18:52,339 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:18:52,340 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:18:52,340 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:18:52,340 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:18:52,340 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:18:52,340 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:18:52,341 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:18:52,341 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 15:18:55,352 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:18:55,353 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:18:55,353 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:18:55,353 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:18:55,354 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:18:55,354 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:18:55,355 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:18:55,355 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 15:18:55,365 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:18:55,366 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:18:55,366 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:18:55,366 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:18:55,367 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:18:55,367 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:18:55,367 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:18:55,367 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 15:35:51,552 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:35:51,552 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:35:51,553 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:35:51,553 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:35:51,554 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:35:51,554 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:35:51,554 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:35:51,555 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
-2025-07-22 15:35:51,561 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor1 smoothed from 11.0000 to 10.5000
-2025-07-22 15:35:51,562 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor1 smoothed from 12.0000 to 11.2500
-2025-07-22 15:35:51,562 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor1 smoothed from 15.0000 to 13.1250
-2025-07-22 15:35:51,562 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor1 smoothed from 15.0000 to 14.0625
-2025-07-22 15:35:51,563 - INFO - Value at timestamp 2023-01-01 00:01:00 in column sensor2 smoothed from 21.0000 to 20.5000
-2025-07-22 15:35:51,563 - INFO - Value at timestamp 2023-01-01 00:02:00 in column sensor2 smoothed from 22.0000 to 21.2500
-2025-07-22 15:35:51,563 - INFO - Value at timestamp 2023-01-01 00:03:00 in column sensor2 smoothed from 23.0000 to 22.1250
-2025-07-22 15:35:51,564 - INFO - Value at timestamp 2023-01-01 00:04:00 in column sensor2 smoothed from 24.0000 to 23.0625
