--- docs/reports/1403-12-15-farid.md
@@ -1,56 +1,0 @@
-### *گزارش کار روزانه*  
-*نام:* فرید کاکی  
-*تاریخ:* ۱۵ اسفند ۱۴۰۲  
-
----
-
-### *۱. مقدمه*  
-در این گزارش، فعالیت‌های انجام‌شده در راستای طراحی و پیاده‌سازی یک *سیستم نظارت، پیش‌بینی و کنترل مبتنی بر هوش مصنوعی* برای صنعت *نفت و گاز* بررسی می‌شود. این سیستم شامل جمع‌آوری داده‌ها، پردازش با مدل‌های یادگیری ماشین و نمایش خروجی از طریق داشبورد است.  
-
----
-
-### *۲. فعالیت‌های انجام‌شده*  
-
-#### *۲.۱ بررسی و انتخاب ابزارهای مانیتورینگ*
-- *Prometheus* برای جمع‌آوری داده‌های زمان‌سری از سنسورها و سیستم‌های کنترلی انتخاب شد.  
-- *Grafana* برای نمایش داده‌های مانیتورینگ در قالب داشبوردهای گرافیکی تست شد.  
-- *Zabbix* به عنوان یک گزینه دیگر برای مانیتورینگ سیستم‌ها ارزیابی شد.  
-- *مقایسه انجام شد:*  
-  - *Prometheus + Grafana* مناسب‌تر برای *تحلیل داده‌های لحظه‌ای و اتصال به یادگیری ماشین*  
-  - *Zabbix* مناسب برای *مدیریت هشدارها و نظارت کلی سیستم*  
-
-#### *۲.۲ بررسی نحوه دریافت داده از Modbus و SCADA*
-- نرم‌افزار *Modbus Slave* برای *شبیه‌سازی داده‌های آزمایشی* راه‌اندازی شد.  
-- امکان ارسال داده‌های *سیمولینک MATLAB* به *Modbus* بررسی شد.  
-- تست اتصال *TIA Portal (Step 7 + WinCC)* برای *دریافت داده از PLC* انجام شد.  
-
-#### *۲.۳ طراحی معماری کلی سیستم*
-- داده‌ها از *PLC (Modbus)* دریافت و در *Prometheus* ذخیره می‌شوند.  
-- مدل‌های *یادگیری ماشین* روی *سرور GPU* پردازش را انجام می‌دهند.  
-- نتایج پردازش از طریق *API در Flask* در اختیار *Grafana / React / LabVIEW* قرار می‌گیرند.  
-- بررسی امکان *پیاده‌سازی یادگیری تقویتی برای تنظیم خودکار پارامترها* (مثل فشار و دما).  
-
-#### *۲.۴ بررسی ابزارهای کنترل و داشبورد*
-- تصمیم‌گیری برای استفاده از *Grafana* برای نمایش داده‌های زنده.  
-- بررسی استفاده از *React / LabVIEW* برای *کنترل مقادیر سنسورها و ارسال فرمان به PLC*.  
-- تست ارسال *داده‌های خروجی مدل هوش مصنوعی* به *Modbus* برای تنظیم خودکار پارامترها.  
-
----
-
-### *۳. چالش‌ها و برنامه‌های آتی*  
-
-#### *۳.۱ چالش‌های پیش‌رو*
-- *بهینه‌سازی ارتباط بین SCADA و Prometheus*  
-- *افزایش دقت پیش‌بینی مدل‌های یادگیری ماشین*  
-- *بررسی نحوه کنترل مستقیم PLC از طریق یادگیری تقویتی*  
-
-#### *۳.۲ برنامه‌های آتی*
-- *اتصال کامل Prometheus به داده‌های واقعی از Modbus*  
-- *ساخت یک مدل یادگیری ماشین برای پیش‌بینی تغییرات فشار و دما*  
-- *ایجاد یک داشبورد با Grafana و React برای نمایش و کنترل داده‌ها*  
-- *یکپارچه‌سازی با Zabbix برای ارسال هشدارهای بحرانی*  
-
----
-
-### *۴. نتیجه‌گیری*
-امروز پیشرفت قابل‌توجهی در انتخاب ابزارهای مناسب برای مانیتورینگ، جمع‌آوری داده و پردازش با یادگیری ماشین حاصل شد. مراحل بعدی شامل *پیاده‌سازی یادگیری تقویتی برای تنظیم پارامترهای PLC و بهینه‌سازی ارتباط بین سیستم‌ها* خواهد بود.
