--- docs/reports/1403-12-15MohammadN.md
@@ -1,12 +1,0 @@
-## گزارش کار محمد نیکوئی
-
-## کار های صورت گرفته
-# 1- بررسی و کسب اطلاعات درباره بازوهای مکانیکی
-# 2- بررسی و کسب اطلاعات در مورد موارد و ویژگی های تاثیرگذار بر سنسور های دستگاه های حفاری نفت و گاز
-# 3- ارائه یک گزارش ساده و کلی درمورد موارد موثر بر اندازه گیری موارد داخلی چاه های نفت و گاز و تهیه یک دیتاست از آن
-# 1- RHOB (Bulk Density)
-# 2- GR (Gamma Ray)
-# 3- DEPTH_MD
-# 4- NPHI (Neutron Porosity)
-# 5- PEF (Photoelectric Factor)
-# 6- DTC (Compressional Wave Velocity)
