--- docs/reports/1403-12-19-Elham.md
@@ -1,18 +1,0 @@
-1403-12-19
-این گزارش شامل مراحل پاک سازی داده برای آماده سازی آنها برای تحلیل و پردازش است. 
-مراحل پاک سازی دادها
-    بارگذاری داده ها 
-    بررسی فرمت داده ها 
-    بررسی حجم داده ها
-بررسی اولیه داده ها
-    بررسی ساختار داده ها 
-    شناسایی ستون های با مقادیر گم شده
-    بررسی داده های تکراری
-اصلاح ناسازگاری ها و خطا ها
-   اصلاح فرمت تاریخ ها و زمان ها
-ذخیره سازی تغییرات اعمال شده
-
-نتایج
-
-داده ها پس از پاک سازی آماده استفاده در مراحل دیگر تحلیل هستند
-
