--- docs/reports/1403-12-20-Elham.md
@@ -1,3 +1,0 @@
-1403-12-20
-شناسایی داده های خارج از محدوده (outliner)
-با استفاده از matplot, seaborn نمودار boxplot رسم کردیم که نشان داده همه داده های ما در محدوده مشخص هستند و داده ای خارج از محدوده نداریم.
