--- docs/reports/1403-12-20MohammadN.md
@@ -1,63 +1,0 @@
-### *گزارش کار روزانه*
-*نام:* محمد نیکوئی  
-*تاریخ:* 20 اسفند 1403  
-
-
-### *1 مقدمه*  
-بررسی *سنسور های موثر بر پایه LWD,MWD در حوزه حفاری چاه های نفت و گاز*
-
-### *2 فعالیت‌های انجام‌شده*
-
-### 2.1 بررسی و جمع آوری داده های موثر بر حفاری نفت و گاز
-
-علاوه بر سنسور های گفته شده. سنسور های مهم دیگری که میتوانند شرایط مختلف دستگاه حفاری را درنظر بگیرند و به کاربر ارائه دهند عبارتند از
-
-- 1- 19- *سنسور های NMR* (ثبت مغناطیس هسته ای- Nuclear Magnetic Resonance Logging)
-    استاندارد های *API RP 40 و ISO 13546*
-  ##  محدوده های مناسب برای هر پارامتر:
-    تخلخل (Porosity): *بین 0 تا 100 درصد*
-    تراوایی (Permeability): *بین0.001 تا 10,000 میلی دارسی (mD)*
-    زمان (Relaxation T1): *0.1 میلی‌ثانیه تا 10000 میلی‌ثانیه*
-    زمان (Relaxation T2): *1 میلی‌ثانیه تا 1,000 میلی‌ثانیه*
-
-
-- 2- 20-*سنسور های مقاومت سنجی* (ثبت مقاومت الکتریکی و القایی -Electrical & Induction Logging)
-
-    استاندارد های *API RP 45 و ISO 13535*
-  ##  محدوده مناسب:
-    مقاومت سازه(مناطق نفتی/گازی): *10 تا 1000 اهم-متر*
-    مقاومت سازه(مناطق آبی): *0.1 تا 10 اهم متر*
-
-- 3- 21-	*سنسور های گاماری* (ثبت طبیعی گاماری- Natural Gamma Ray Logging)
-    استاندارد های* API RP 32 و ISO 18589*
-  ##  محدوده های گاماری:
-    تشخیص گاماری: *0 تا 300 واحد* API (محدوده معمول)
-    تشخیص شیل: *بیشتر از 75 واحد* API (گامای بالا)
-    تشخیص ماسه سنگ: *کمتر از 50 واحد* API (گامای پایین)
-
-
-- 4- 22-*سنسور چگالی*(ثبت چگالی حجمی – Bulk Density Logging)
-استاندارد های *API RP 39 و ISO 13500*
-  ##  محدوده های مناسب
-چگالی حجمی: *1.5 تا 2.95 گرم بر سانتیمتر مکعب*
-
-- 5-*سنسور نوترونی* (ثبت تخلخل نوترونی – Neutron Porosity Logging): مناسب برای شناسایی تخلخل یا همان Porosity
-    استاندارد های *API RP 41 و ISO 11961*
-  ##  محدوده مناسب تخلخل:
-    0 تا 100 درصد
-
-### موانع و مشکلات
-
-## عدم توانایی در پیدا کردن دیتاست مناسب
-
-*سنسورهای نوری (Optical Sensors)* به دلیل دقت بالا، مقاومت در برابر شرایط سخت محیطی و قابلیت ارائه داده‌های بلادرنگ، به یکی از فناوری‌های کلیدی در صنعت نفت و گاز تبدیل شده‌اند. این سنسورها در کاربردهایی مانند نظارت بر چاه‌های نفت، تشخیص نشت در خطوط لوله و آنالیز ترکیبات شیمیایی مورد استفاده قرار می‌گیرند. با این حال، یکی از چالش‌های اصلی در توسعه و بهینه‌سازی این سنسورها، عدم وجود دیتاست‌های مناسب و در دسترس است.
-
-### چالش‌ها و موانع
-## عدم دسترسی به دیتاست‌های عمومی
-
-بسیاری از داده‌های مربوط به سنسورهای نوری در صنعت نفت و گاز، به دلیل مسائل امنیتی و رقابتی، به صورت عمومی منتشر نمی‌شوند.
-شرکت‌های بزرگ نفتی و گازی معمولاً داده‌های خود را به عنوان اطلاعات محرمانه در نظر می‌گیرند و آن‌ها را در اختیار محققان قرار نمی‌دهند.
-
-## محدودیت در دیتاست‌های موجود
-
-دیتاست‌های موجود اغلب عمومی و کلی هستند و به طور خاص برای سنسورهای نوری در محیط‌های نفتی طراحی نشده‌اند.
