--- docs/reports/1403-12-22MohammadN.md
@@ -1,25 +1,0 @@
-### *گزارش کار روزانه*
-*نام:* محمد نیکوئی  
-*تاریخ:* 22 اسفند 1403  
-
-
-### *1 مقدمه* 
-*انجام تمرین* و بررسی برای پیدا کردن دیتاست برای *سنسور های نوری*
-
-### *2 فعالیت‌های انجام‌شده*
-## 2.1 دیدن ویدیو های آموزشی و تمرین برای *Git&Github*
-
-ویدیو ها از سایت دانلودلی *www.downloadly.ir* گردآوری شده اند
-*Programming with mosh:Git and Github*
-
-### 2.2 بررسی هرچه بیشتر بابت پیدا کردن دیتاست برای سنسور های نوری(Optic Sensors) در حوزه حفاری نفت و گاز
-   ## چالش‌ها و موانع
-   *عدم توانایی در پیدا و جمع آوری کردن دیتاست برای سنسور های نوری*
-### تاثیرات سنسور های نوری در حوزه حفاری چاه های نفت و گاز 
-## نظارت بر چاه‌های نفت و گاز
-- 1- *اندازه گیری دما و فشار*
-- 2- *جلوگیری از نشتی در لوله ها*
-- 3- *پایش شرایط چاه در طول فرآیند حفاری و بهبود بخشدین به عملیات حفاری*
-- 4- *بهبود و افزایش ایمنی کارکنان و چاه ها*
-- 5- *کاربرد در فناوری دوقلوی دیجیتال(Digital Twin)*
-# Note: بدلیل تاثیرات هرچه بیشتر و تمامتر این ویژگی در حوزه حفاری نفت و گاز، این ویژگی به یکی از مهمترین ویژگی های Drilling Automation & Digital Twins بدل شده و نیاز است که به جهت پردازش و افزایش سرعت در انجام پروژه، این دیتاست هرچه سریع تر در اختیار شرکت قرار گیرد.
