--- docs/reports/1403-12-25MohammadN.md
@@ -1,23 +1,0 @@
-### *گزارش کار روزانه*
-*نام:* محمد نیکوئی  
-*تاریخ:* 25 اسفند 1403  
-
-
-### *1 مقدمه*  
-*بررسی داده های جمع آوری شده برای حفاری چاه های نفت و گاز و نرمالایز و تمیز کردن داده*
-
-
-### *2 فعالیت‌های انجام‌شده*
-## *2.1 نرمال کردن و تمیز کردن داده*
-- 1- *در ابتدا مقادیر NaN و null و Outlier از مقادیر دیتاست حذف شد*
-- *لازم به ذکر است که دیتاست بدست آمده مقادیر outlier و null زیادی را دارا نبود و تنها 10 درصد از دیتاست شامل مقادیر outlier و null بود*
-
-
-
-## *2.2 تماشای ویدیو آموزشی* 
-- *دوره آموزشی دیده شده: (دوره Daniel Bourke-سایت Downloadly.com-دوره ی Complete A.I. & Machine Learning, Data Science Bootcamp)* 
-
-- *موارد دیده شده: Analyse data with Logistic Regression and Knn Models in python*
-
-### 3 *برنامه های آتی*
-*پیاده سازی مدل های xgboost,RandomForest,TPOT بر روی داده ها و دریافت خروجی onnxfile*
