--- docs/reports/1404-01-16-Elham.md
@@ -1,18 +1,0 @@
-:فعالیت‌های انجام‌شده
-
-:MySQL ایجاد پایگاه داده در 
-ایجاد شد my sqlیک دیتابیس برای داده‌های جمع‌آوری‌شده از سنسورها طراحی و در
- این پایگاه داده شامل جداولی برای ذخیره‌سازی مقادیر سنسورها به همراه زمان ثبت آن‌ها است
-
-grafana راه‌اندازی داشبورد در
-داشبورد اولیه جهت مشاهده داده‌های سنسورها ایجادشد  MySQL به Grafana با اتصال
-
-:مطالعه ویدیوهای آموزشی
-جهت تسلط بیشتر بر فرآیندها، ویدیوهای آموزشی مربوط به لینوکس و ابزارهای 
-گرافانا مشاهده شد این آموزش‌ها شامل مفاهیم پایه‌ای سیستم‌عامل لینوکس و نحوه مانیتورینگ با گرافانا بودند که به درک بهتر محیط کاری کمک کردند
-
-:چالش ها
-در حال حاضر، داده‌ها به صورت استریمدر داشبورد نمایش داده نمی‌شوند. این مشکل احتمالاً ناشی از عدم پیکربندی صحیح منبع داده یا محدودیت‌های ارتباطی با دیتابیس است که نیاز به بررسی بیشتر دارد.
-
-:گام‌های بعدی
-بررسی اتصال زنده دیتابیس با گرافانا جهت فعال‌سازی نمایش داده‌های لحظه‌ای
