--- docs/reports/1404-01-16-MohammadN.md
@@ -1,45 +1,0 @@
-### *گزارش کار روزانه*  
-*نام:* محمد نیکوئی 
-*تاریخ:* 1404 فروردین 16
-
----
-
-### *۱. مقدمه*  
-در این گزارش، فعالیت‌های انجام‌شده در راستای طراحی و پیاده‌سازی *روش پیشرفته برای توسعه مدل‌های پیش‌بینی در حفاری* به انجام رسیده است 
----
-
-### *۲. فعالیت‌های انجام‌شده* 
-## 2.1 تکمیل و چمع آوری داده های مربوط به *I DRILL*
-یک دیتاست جدید شامل ده میلیون داده جدید با وجود ویژگی های 
-* 1   Pressure Sensors (PSI)                10000000 non-null  float64
-* 2   Temperature Sensors (Celsius)         10000000 non-null  float64
-* 3   Flow Sensors (GPM)                    10000000 non-null  float64
-* 4   Vibration Sensors (mm/s)              10000000 non-null  float64
-* 5   Gas Sensors (PPM)                     10000000 non-null  float64
-* 6   Position Sensors (Degrees)            10000000 non-null  float64
-* 7   Tension Sensors (Tons)                10000000 non-null  float64
-* 8   Acoustic Sensors (Hz)                 10000000 non-null  float64
-* 9   Humidity Sensors (% RH)               10000000 non-null  float64
-* 10  Optical Sensors (nm)                  10000000 non-null  float64
-* 11  Magnetic Sensors (Tesla)              10000000 non-null  float64
-* 12  Seismic Sensors (g)                   10000000 non-null  float64
-* 13  Chemical Sensors (% Concentration)    10000000 non-null  float64
-* 14  ROP Sensors (Feet/Hour)               10000000 non-null  float64
-* 15  WOB Sensors (Pounds)                  10000000 non-null  float64
-* 16  RPM Sensors (RPM)                     10000000 non-null  float64
-* 17  Leak Sensors (Liters/hour)            10000000 non-null  float64
-* 18  Corrosion Sensors (mm/year)           10000000 non-null  float64
-* 19  NMR Porosity (%)                      10000000 non-null  float64
-* 20  NMR Permeability (mD)                 10000000 non-null  float64
-* 21  NMR T1 (ms)                           10000000 non-null  float64
-* 22  NMR T2 (ms)                           10000000 non-null  float64
-* 23  Electrical Logging (Oil/Gas) (Ohm-m)  10000000 non-null  float64
-* 24  Electrical Logging (Water) (Ohm-m)    10000000 non-null  float64
-* 25  Gamma Ray Logging (API)               10000000 non-null  float64
-* 26  Density Logging (g/cm^3)              10000000 non-null  float64
-* 27  Neutron Porosity (%)                  10000000 non-null  float64
-شکل گرفت که در بخش *dataset* قرار گرفت
-
-
-https://limewire.com/?referrer=pq7i8xx7p2
-*لینک دیتاست*
