--- docs/reports/1404-01-17-Elham.md
@@ -1,17 +1,0 @@
-گزارش کار: مانیتورینگ سنسورها با Prometheus و Grafana
-
-مراحل انجام کار:
-
-مشاهده ویدیو آموزشی:
-ویدیوهای مرتبط با نصب، پیکربندی و اتصال Prometheus به Grafana دیده شد.
-
-اتصال Prometheus به Grafana:
-Prometheus به عنوان منبع داده در Grafana تنظیم شد و اتصال با موفقیت برقرار شد.
-
-
-دسته‌بندی سنسورها:
-سنسورها به ۷ گروه تقسیم شدند:
-
-
-ایجاد داشبورد برای هر گروه:
-برای هر دسته داشبورد جداگانه طراحی شد که شامل نمودارهای زمانی، مقدار لحظه‌ای، هشدارها و جداول داده‌ها بود.
