--- docs/reports/1404-01-20MohammadN.md
@@ -1,16 +1,0 @@
-### *گزارش کار روزانه*
-*نام:* محمد نیکوئی  
-*تاریخ:* 20 فروردین 1404  
-
-
-### *1 مقدمه*  
-*برسی و پیاده سازی مدل های سری زمانی و سمجش دقت ان ها برای دیبتاست *
-
-
-### *2 فعالیت‌های انجام‌شده*
-مدل های سری زمانی *LSTM* از طریق ماژول پاینورچ پیاده سازی شد و دقت کمتر از 4 درصد به همراه حاشیه خطای بالای 50 خروجی داده شد 
-
-
-### *3 چالش ها *  
-1 عدم وجود همبستگی (CORRELATION) میان ویژگی های موجود در دیتاست 
-عدم وجود داده های واقعی 2
