--- docs/reports/1404-01-23-Elham.md
@@ -1,8 +1,0 @@
-اتصال Windows Exporter به Prometheus:
-طی روزهای اخیر، اتصال Windows Exporter به Prometheus که با باگ همراه بود، بررسی و اشکال‌زدایی شد. پس از شناسایی منشأ خطا و اعمال تنظیمات لازم در فایل پیکربندی Prometheus و بررسی وضعیت سرویس Windows Exporter، اتصال پایدار برقرار شد و داده‌ها با موفقیت در Prometheus جمع‌آوری و مانیتور شدند.
-
-ساخت داشبوردهای نمونه برای Dataset:
-به منظور تحلیل داده‌های جمع‌آوری‌شده از Windows Exporter، چند داشبورد نمونه در Grafana طراحی شد و داده‌ها به صورت گرافیکی نمایش داده شدند تا امکان مانیتورینگ فراهم شود.
-
-مطالعه ویدیوی آموزشی درباره لینوکس و Grafana:
-یک ویدیوی آموزشی درباره اصول پایه‌ای سیستم‌عامل لینوکس و همچنین آشنایی با قابلیت‌های کلیدی Grafana مشاهده شد که به درک بهتر ابزارهای مانیتورینگ کمک کرد.
