--- docs/reports/1404-01-23MohammadN.md
@@ -1,22 +1,0 @@
-### *گزارش کار روزانه*
-*نام:* محمد نیکوئی  
-*تاریخ:* 23 فروردین 1404  
-
-
-### *1 مقدمه*  
-*برسی و پیاده سازی مدل های سری زمانی و سمجش دقت ان ها برای دیبتاست به توسط ماژول *darts* *
-
-
-### *2 فعالیت‌های انجام‌شده*
-* مدل های سری زمانی *LSTM* از طریق ماژول دارتس پیاده سازی شد و دقت کمتر از 4 درصد به همراه حاشیه خطای بالای 50 خروجی داده شد 
-* پیاده سازی مدل *LinearRegression* بر روی دبتاست موجود باعث پدیددامدن دقت بالای 99 درصد و حاشیه خطای کمتر از 0.27 
-
-
-### *3 چالش ها *  
-1 عدم وجود همبستگی (CORRELATION) میان ویژگی های موجود در دیتاست 
-عدم وجود داده های واقعی 2
-عدم دریافت خروجی مناسب از طریق مدل های ساخته شده در ماژول دارتس 
-
-
-### *4 گام های بعدی*  
-پیاده سازی مدل های *xgboost* & *catboost* بر روی دیتاست موجود
