--- docs/reports/1404-01-24-Elham.md
@@ -1,1 +1,0 @@
- یک Exporter سفارشی برای دریافت داده‌های سنسورها طراحی شد. این Exporter داده‌ها را از طریق پروتکل مشخص (مثلاً HTTP یا Serial) جمع‌آوری کرده و آن‌ها را در قالب استاندارد Prometheus (متریک‌های متنی) در اختیار قرار می‌دهد. پس از راه‌اندازی Exporter، Prometheus به‌طور دوره‌ای داده‌ها را از آن جمع‌آوری می‌کند. در ادامه، از طریق اتصال Prometheus به Grafana، داشبوردی برای نمایش بلادرنگ مقادیر سنسورها طراحی شد
