--- docs/reports/1404-01-25MohammadN.md
@@ -1,15 +1,0 @@
-### *گزارش کار روزانه*
-*نام:* محمد نیکوئی  
-*تاریخ:* 25 فروردین 1404  
-
-
-### *1 مقدمه*  
-*گار بر روی پروپوزال*
-
-### *2 فعالیت‌های انجام‌شده*
-* نوشتن و کار بر روی 
-*طراحی سامانه هوشمند برنامه‌ریزی و بهینه‌سازی تولید گاز مایع با تکیه بر مدل‌سازی ریاضی و الگوریتم‌های پیشرفته هوش مصنوعی در پالایشگاه آبادان* 
-
-
-### *4 گام های بعدی*  
-*کار بیشتر بر روی پروپوزال و ادامه و اتمام آن*
