--- docs/reports/1404-01-27MohammadN.md
@@ -1,17 +1,0 @@
-### *گزارش کار روزانه*
-*نام:* محمد نیکوئی  
-*تاریخ:* 27 فروردین 1404  
-
-
-### *1 مقدمه*  
-*گار بر روی پروپوزال*
-
-### *2 فعالیت‌های انجام‌شده*
-
-* 1-کار بر روی پروپوزال
-* نوشتن و کار بر روی 
-*طراحی سامانه هوشمند برنامه‌ریزی و بهینه‌سازی تولید گاز مایع با تکیه بر مدل‌سازی ریاضی و الگوریتم‌های پیشرفته هوش مصنوعی در پالایشگاه آبادان* 
-
-### *4 گام های بعدی*  
-*کار بر روی ISSUE11*
-*روش پیشرفته برای توسعه مدل‌های پیش‌بینی در حفاری*
