--- docs/reports/1404-01-30-Elham.md
@@ -1,32 +1,0 @@
-تاریخ: ۱۴۰۳/۰۱/۳۰
-الهام آرمین
-
-۱. مقدمه
-هدف امروز بررسی نحوه اتصال داده‌های سفارشی (custom metrics) به ابزار مانیتورینگ Grafana از طریق Prometheus بود. برای این منظور، یک Exporter سفارشی طراحی شد تا داده‌ها را در قالب قابل خواندن توسط Prometheus فراهم کند.
-
-۲. ابزارها و تکنولوژی‌های استفاده‌شده
-
-Prometheus: برای جمع‌آوری داده‌های مانیتورینگ
-
-Grafana: برای نمایش گرافیکی داده‌ها
-
-Custom Exporter: نوشته‌شده با Python 
-
-HTTP endpoint: برای ارائه داده‌ها از Exporter
-
-۳. مراحل اجرا
-توسعه Exporter برای تولید متریک‌های سفارشی با استفاده از فرمت استاندارد Prometheus
-
-راه‌اندازی Prometheus و تنظیم فایل prometheus.yml برای scrape کردن Exporter:
-
-scrape_configs:
-  - job_name: 'custom_exporter'
-    static_configs:
-      - targets: ['localhost:8000']
-  
-  
-  راه‌اندازی Grafana و افزودن Prometheus به‌عنوان data source
-تلاش برای ایجاد داشبورد در Grafana بر اساس متریک‌های جمع‌آوری‌شده
-
-۴. مشکل به‌وجودآمده
-با وجود اینکه Prometheus به درستی داده‌ها را از Exporter دریافت می‌کرد (تأیید شده با آدرس http://localhost:9090/targets و http://localhost:9090/metrics)، اما متریک‌ها در Grafana نمایش داده نمی‌شدند
