--- docs/reports/1404-01-30MohammadN.md
@@ -1,33 +1,0 @@
-
-# گزارش کار روزانه  
-**نام:** محمد نیکوئی  
-**تاریخ:** 30 فروردین 1404  
-
----
-
-## 1. مقدمه  
-امروز بخشی از زمان خود را به یادگیری مفاهیم پایه‌ای و مهم در حوزه یادگیری ماشین (Machine Learning) اختصاص دادم. تمرکز اصلی بر روی تکنیک‌های پیش‌پردازش داده‌ها بود که یکی از مراحل کلیدی در آماده‌سازی داده برای مدل‌سازی به‌شمار می‌رود.
-
----
-
-## 2. فعالیت‌های انجام‌شده  
-- مطالعه و پیاده‌سازی روش **One-Hot Encoding** برای تبدیل ویژگی‌های دسته‌ای (Categorical) به قالب عددی که برای مدل‌های یادگیری ماشین قابل درک باشند.  
-- بررسی و استفاده از روش‌های **پر کردن مقادیر گمشده (Filling Missing Values)** با استفاده از میانگین، میانه یا مقدار ثابت.  
-- تمرین بر روی حذف مقادیر گمشده با استفاده از **Drop NaN Values**، و بررسی تاثیر آن بر کیفیت و حجم داده‌ها.  
-- استفاده از پکیج‌های مرتبط در پایتون نظیر Pandas برای پیاده‌سازی این تکنیک‌ها به صورت عملی.
-
----
-
-## 3. چالش‌ها  
-- درک تفاوت‌های ظریف بین روش‌های پر کردن مقادیر گمشده و انتخاب بهترین روش برای هر نوع داده.  
-- مواجهه با دیتاست‌هایی که ترکیبی از داده‌های عددی و متنی داشتند و نیاز به دقت بیشتر در انتخاب روش مناسب برای پیش‌پردازش.  
-- نیاز به تمرین بیشتر در پیاده‌سازی کدهای تمیز و قابل فهم برای این فرآیندها.
-
----
-
-## 4. گام‌های بعدی  
-- ادامه تمرین بر روی دیتاست‌های واقعی و بزرگ‌تر برای تسلط بیشتر بر روش‌های پیش‌پردازش.  
-- یادگیری تکنیک‌های دیگر مانند **Normalization** و **Standardization** برای آماده‌سازی بهتر داده‌ها.  
-- مطالعه مفاهیم اولیه مدل‌های یادگیری نظارتی مانند **Linear Regression** و **Decision Trees** به عنوان قدم بعدی در مسیر یادگیری ماشین.
-
----
