--- docs/reports/1404-01-31-Elham.md
@@ -1,10 +1,0 @@
-گزارش عملکرد روزانه
-تاریخ: ۱۴۰۴/۰۱/۳۱
-نام: الهام آرمین
-
-امروز نرم‌افزار InfluxDB با موفقیت نصب شد. هدف از این مرحله، فراهم‌سازی زیرساخت مناسب برای ذخیره و پایش داده‌های سری زمانی بود.
-با این حال، در مرحله‌ی اتصال Telegraf به InfluxDB جهت انتقال داده‌ها، با مشکل مواجه شدم و سیستم نتوانست اتصال موفقیت‌آمیزی بین Telegraf و InfluxDB برقرار کند. در ادامه، قصد دارم با بررسی فایل‌های پیکربندی Telegraf و لاگ‌های مربوطه، دلیل این عدم اتصال را شناسایی و برطرف کنم.
-
-از سوی دیگر، با توجه به اینکه Kafka با موفقیت توانسته داده‌ها را به پایگاه داده  منتقل کند، اکنون تمرکز بر روی نمایش داده‌ها به صورت استریم و بلادرنگ (real-time) در Grafana است. برای دستیابی به این هدف، بهترین راهکار فنی و زیرساختی، استفاده از یک سرور مجازی می‌باشد.
-
-با تهیه‌ی سرور مجازی، امکان پیاده‌سازی و انتشار یک HTTP endpoint فراهم می‌شود. این سرویس قادر خواهد بود داده‌های پردازش‌شده را از Kafka یا InfluxDB دریافت کرده و از طریق یک API قابل دسترس، در اختیار Grafana قرار دهد. این راهکار، همکاری تیمی، مانیتورینگ و تحلیل داده‌ها را تسهیل کرده و مقیاس‌پذیری سیستم را نیز افزایش خواهد داد.
