--- docs/reports/1404-4-31.md
@@ -1,11 +1,0 @@
-گزارش روزانه
-نام و نام خانوادگی:سجاد طهماسبی
-تاریخ:1404-4-31
-تاریخ شروع همکاری:1404-4-22
-نوع همکاری:کاراموزی
-فعالیت ها:انجام تسک Data Imputation and Reconciliation
-ادامه تمرین رگراسیون خطی و مهارت افزایی دیتا و ماشین لرنینگ همراه با کتاب 
-Hands on Machine Learning
-اهداف فردا:
-انجام تسک بعدی React  در صورت امکان
-و ادامه ی مهارت افزایی با کتاب
