--- docs/reports/report.txt
@@ -1,43 +1,0 @@
-این دستورالعمل برای تهیه و ارسال گزارش روزانه کاری بسیار سازمان‌یافته و دقیق است. در ادامه، به بررسی و خلاصه‌ای از محتوای آن می‌پردازیم:
-
----
-
-## **خلاصه دستورالعمل نوشتن و ارسال گزارش روزانه**
-
-### **1. قالب گزارش**
-- **تاریخ:** به فرمت YYYY/MM/DD  
-- **نام و نقش:** ذکر نام و نقش شما در پروژه  
-- **کارهای انجام‌شده:** لیستی از فعالیت‌های انجام‌شده با توضیح کوتاه  
-- **چالش‌ها و مشکلات:** ذکر مشکلات و موانع موجود  
-- **برنامه برای فردا:** لیستی از کارهایی که قرار است انجام شود  
-- **پیشنهادات (اختیاری):** ارائه پیشنهادات برای بهبود کار  
-
----
-
-### **2. مثال گزارش**
-یک مثال عملی از گزارش روزانه ارائه شده است که شامل:
-- تست‌های واحد، بهینه‌سازی API، و به‌روزرسانی مستندات  
-- مشکل در اتصال به پایگاه داده  
-- برنامه‌ریزی برای رفع مشکلات و انجام تست‌های انتگریشن  
-- پیشنهاد استفاده از ابزار جدید مدیریت وظایف  
-
----
-
-### **3. نحوه ارسال**
-- **فرمت فایل:** `.md` (Markdown)  
-- **نام فایل:** `نام_خود_تاریخ.md` (مثال: `AliAhmadi_20231010.md`)  
-- **مهلت ارسال:** حداکثر ساعت 18:00 هر روز  
-- **کانال ارسال:** ایمیل یا کانال ارتباطی تعیین‌شده توسط مدیر پروژه  
-
----
-
-### **4. نکات مهم**
-- **دقت و شفافیت:** تمام فعالیت‌ها و مشکلات باید به‌طور واضح و دقیق توضیح داده شوند.  
-- **رعایت زمان:** ارسال گزارش در پایان هر روز الزامی است.  
-- **شفافیت در مشکلات:** مشکلات باید به‌صورت صریح ذکر شوند تا مدیر پروژه بتواند راه‌حل مناسب ارائه دهد.  
-
----
-
-این دستورالعمل به شما کمک می‌کند تا گزارش‌های روزانه خود را به‌صورت استاندارد و مؤثر تهیه کنید و به مدیریت پروژه کمک کنید تا پیشرفت کار را به‌درستی پیگیری کند.  
-
-اگر سوالی دارید یا نیاز به توضیحات بیشتری دارید، خوشحال می‌شوم کمک کنم! 😊
