--- docs/reports/sajjad tahmasebi dailyreports/1404-4-24.md
@@ -1,16 +1,0 @@
-گزارش روزانه
-نام و نام خانوادگی:سجاد طهماسبی
-تیم:i_drill
-تاریخ شروع همکاری:1404-4-22
-تاریخ گزارش:1404-4-24
-نوع همکاری:کاراموز
-فعالیت های امروز:تمرین پایتون
-تمرین ریاضیات ام ال
-زمان فعالیت:4 ساعت 
-اهداف فردا:تمرین پنداز و نامپای
-شروع اولین تسک با عنوان ام ال کلاسیک و بهینه سازی پارامتر ها 
-یادگیری مفهوم های یادگیری نظارت شده و نظارت نشده و الگوریتم ها
-یادگیری کد زنی با scikit-learn
-تسک هایی که به من تعلق دارند:یادگیری ام ال کلاسیک و بهینه سازی پارامترها
-یادگیری تقویتی برای بهینه سازی
-روش پیشرفته برای توسعه مدل های پیش بینی در حفاری
