--- docs/reports/sajjad tahmasebi dailyreports/1404-4-25
@@ -1,19 +1,0 @@
-گزارش روزانه
-نام و نام خانوادگی:سجاد طهماسبی
-نوع همکاری:کاراموزی
-تاریخ شروع همکاری:1404-4-22
-تاریخ گزارش:1404-4-25
-فعالیت های انجام شده:تمرین ریاضیات
-تمرین پایتون
-تمرین کتابخانه های نامپای و پنداز
- تسک یادگیری ماشین کلاسیک و بهینه سازی پارامتر ها
-منابع یادگیری: دوره اندرو ان جی برای مفاهیم
-کتاب hands on machine learning 
-ویدیو های یوتیوب برای کد زنی scikit learn 
-اهداف فردا :تمرکز بیشتر روی تسک یادگیری ماشین و بهینه سازی پارامتر هاروش یادگیری و مهارت افزایی:
-ابتدا مفاهیم و الگوریتم ها را یاد گرفته سپس کد زنی رو شروع میکنیم
-برای مثال هنگام یاد گیری رگراسیون لاجیستیک ابتدا مفاهیم رو یاد میگیریم و سپس کد زنی
-به نوعی تایمی که هنوز برای هیچ یک از تسک ها اعلام نشده به 3 بخش تقسیم میشه
-بخش1=یادگیری الگوریتم ها و مفاهیم ریاضی
-بخش2=کدزنی
-بخش3=انجام تسک
