--- docs/reports/sajjad tahmasebi dailyreports/1404-4-29.md
@@ -1,9 +1,0 @@
-گزارش روزانه
-تاریخ:۱۴۰۴،۴،۲۹
-تاریخ شروع همکاری:۱۴۰۴،۴،۲۲
-نوع همکاری:کارآموز
-فعالیت ها:یادگیری ابزارهای تحلیل داده به طور کامل از جمله Pandas ،Numpy، Matplotlib 
-تمرین ریاضیات
-یادگیری مدل های پایه ای کلاسبو ار جمله رگراسیون خطی،تابع هزینه و مفهوم گرادیان کاهشی
-مباحث آماری از جملهstd،capped data،correlation
-اهداف فردا:ادامه ی کتاب hands on machine learning
