--- docs/reports/sajjad tahmasebi dailyreports/1404-5-26.md
@@ -1,11 +1,0 @@
-گزارش روزانه 
-نام و نام خانوادگی:سجاد طهماسبی 
-نوع همکاری:کاراموزی
-زمان شروع همکاری:1404-4-22
-تاریخ گزارش:1404-4-26
-فعالیت های انجام شده:یادگیری پنداز
-تمرین پایتون و نامپای و پنداز
-یادگیری ام ال و شروع مهارت افزایی برای تسک ام ال کلاسیک و بهینه سازی پارامترها
-اهداف فردا:تمرین پایتون
-تمرین ریاضی
-ادامه یادگیری ام ال
