--- docs/reports/sajjad tahmasebi dailyreports/1404.4.28.md
@@ -1,11 +1,0 @@
-گزارش روزانه
-نام و نام خانوادگی:سجاد طهماسبی
-نوع فعالیت:کاراموزی
-تاریخ شروع همکاری:1404.4.22
-تاریخ گزارش:1404.4.28
-فعالیت:ریاضیات ام ال
-یادگیری پایه های ام ال کلاسیک
-تمرین پایتون و کتابخانه های تحلیل داده
-زمان:3 و نیم ساعت
-اعداف فردا:ادامه ی یادگیری ریاضیات ام ال
-تمرین پنداز
